added slideshow
