--- src/movieData.js
@@ -1,53 +1,0 @@
-export const homeData = [
-    {
-      id: 1,
-      name: "Sand Dust",
-      rating: 4.7,
-      time: "2hr : 22mins",
-      desc: "Sand and dust storms (SDS), also known as sirocco, haboob, yellow dust, white storms, and the harmattan, are a natural phenomenon linked with land and water management and climate change.",
-      starring: "Karen Gilchrist, James Earl Jones",
-      genres: "Action",
-      tags: "Action, Adventures, Horror",
-      cover: "../images/home1.jpg",
-      video: "../video/video1.mp4",
-      date: "20-Jan-1997",
-    },
-    {
-      id: 2,
-      // name: "The Mabaso's Family Reunion",
-      rating: 4.6,
-      time: "2hr : 22mins",
-      desc: "The island, like its inhabitants, is wholly fictional. Kong's island home in the Pacific makes its first appearance, along with Kong himself, in the 1933 film King Kong. ",
-      starring: "Brenda Chapman, Jeff Nathanson",
-      genres: "Adventures",
-      tags: "Adventures,Animation,Family",
-      cover: "../images/testing.jpg",
-      date: "10-JUL-2021",
-    },
-    {
-      id: 3,
-      name: "Paul vs Tyson",
-      rating: 4.8,
-      time: "4hr : 43mins",
-      desc: "Jake paul battles mike Tyson as they headline this must-see boxing mega-event from AT&T Stadium in Arlington,Taxes ",
-      starring: "Mike Tyson, jake Paul< Katie Taylor",
-      genres: "Boxing,Sports, Events",
-      tags: "Adventures, Action ",
-      cover: "../images/1.jpg",
-    
-      date: "16 Novermber 2024",
-    },
-    {
-      id: 4,
-      name: "Sand Dust",
-      rating: 4.9,
-      time: "2hr : 22mins",
-      desc: "Sand and dust storms (SDS), also known as sirocco, haboob, yellow dust, white storms, and the harmattan, are a natural phenomenon linked with land and water management and climate change.",
-      starring: "Karen Gilchrist, James Earl Jones",
-      genres: "Action",
-      tags: "Action, Adventures, Horror",
-      cover: "../images/home4.jpg",
-      video: "../video/video4.mp4",
-      date: "12-Aug-2015",
-    },
-  ]
